win32: exclude broken tests
The ceph_test_rados_striper_api_aio tests seem to use the
AioTestData semaphore incorrectly. The same AioTestData structure
is reused for multiple async callbacks, which implies multiple
completion notifications.

However, the semaphore is initialized with 0, so "wait" will
return before subsequent "notify" calls complete. For this reason,
we can have situations in which "notify" callbacks occur after the
test completes and the AioTestData structure and semaphore are
cleaned up, which leads to a crash:

  * https://pastebin.com/raw/gh7CytHA
  * https://pastebin.com/raw/2Yu90uEG

For now, we'll skip those tests.