allow running the same test with different arguments
Some of the functional tests receive various arguments. In some
cases we may want to run the same test multiple times, with
different arguments. For example, we may want to run the librbd
fsx test with multiple backends (e.g. librbd and then rbd-wnbd).

For this reason, we'll convert our string test arguments to
hashtables. We'll use hashtables instead of arrays so that we
may associate test name suffixes with every set of arguments.

For example, ceph_test_librbd_fsx may have the following
subtests:

  * ceph_test_librbd_fsx.librbd
  * ceph_test_librbd_fsx.rbd-wnbd

What's nice about using "." as a suffix separator is that the
subunit parser will group those tests under a single suite.

Signed-off-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>