Temporarily disable perf counter tests
The perf counter tests were already disabled, known to throw
WSAECONNREFUSED admin socket errors when running under Powershell
jobs.

We got similar issues with unittest_perf_counters_cache.exe and
assumed that simply running them in isolated mode would solve the
issue, however it did not.

We'll need to further investigate how Powershell jobs affect
those tests. In the meantime, we'll add them to the "excluded" list
in order to unblock the CI and allow the according PR to merge.

Signed-off-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>