Skip LibRadosMiscConnectFailure.ConnectTimeout
This test sets a connection timeout of 2s and asserts that the
connection attempt completes in less than 4s.

On Windows, it sometimes takes between 5 and 6 seconds, leading to
a test failure. This could be caused by reduced clock or timer
precision.

    RUN      ] LibRadosMiscConnectFailure.ConnectTimeout
    /home/ubuntu/ceph/src/test/librados/misc.cc:87: Failure
    Expected: (dur) < (utime_t(4, 0)), actual: 5.987712 vs 4.000000

We're going to skip this test for the time being.

Signed-off-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>